adds correct rotation for corners on main image track definition

+++ src/main/java/uk/co/mruoc/race/gui/AngledPoint.java
@@ -4,14 +4,14 @@ import java.awt.*;
 
 public class AngledPoint extends Point {
 
-    private final int angle;
+    private final double angle;
 
-    public AngledPoint(int x, int y, int angle) {
+    public AngledPoint(int x, int y, double angle) {
         super(x, y);
         this.angle = angle;
     }
 
-    public int getAngle() {
+    public double getAngle() {
         return angle;
     }
